pyproject.toml
[project]
name = "video-agents"
version = "0.1.0"
description = "MCP Server for automated video and audio editing with AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "faster-whisper>=1.2.0",
    "silero-vad>=5.1.0",
    "scenedetect[opencv-headless]>=0.6.4",
    "ffmpeg-python>=0.2.0",
    "fastmcp[tasks]",
    "toon-format @ git+https://github.com/toon-format/toon-python.git",
    "pytest",
    "torch>=2.4.0",
    "onnxruntime>=1.24.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest-cov>=7.0.0",
]

[project.scripts]
video-agents = "src.server:mcp.run"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise",
]

# CUDA indexing removed for CPU-only testing as requested by user.
# For production GPU usage, refer to research.md.
